+++ CHANGELOG.md
@@ -69,7 +69,7 @@ All notable changes to this project will be documented in this file.
 
 - The following **experimental** factory functions (issue [#258]):
   - `toNonZeroIntOrNull` and `toNonZeroIntOrThrow`
-  - `toPositiveIntOrNull`
+  - `toPositiveIntOrNull` and `toPositiveIntOrThrow`.
 
 [#258]: https://github.com/kotools/types/issues/258